==> 01_Pobieranie_danych.ipynb <==
 # Opis notatnika

 Ten notatnik inicjuje naszą pracę nad warsztatem końcowym. Naszym zadaniem tutaj jest pobranie udostępnionych nam danych do obszaru roboczego, które w następnym kroku wgramy na naszą bazę danych. Ich obróbka oraz analiza zostanie przeprowadzona w specjalnie do tego celu przygotowanych kolejnych notatnikach.

 Na potrzeby tego warsztatu został stworzony dedykowany serwis API, który dostępny jest pod adresem: https://api-datalab.coderslab.com/api/v2. Dodatkowo udostępniona została dokumentacja, z którą można zapoznać się tutaj: [klik](https://api-datalab.coderslab.com/v2/docs/).

 > Dokumentacja jest czysto techniczna i ma na celu prezentację dostępnych endpointów wraz ze zwracanym typem. W celu przetestowania należy kliknąć przysisk `Authorize`, podać token (dostępny poniżej), a następnie `Try it out!` oraz uzupełnić wymagane pola (parametry requesta).

 Zgodnie z dokumentacją stwierdzamy, że udostępnione zostały nam 4 endpointy:
 - `airport` - dane o lotnisku,
 - `weather` - informacje o zarejestrowaniej pogodzie na lotnisku danego dnia,
 - `aircraft` - dane o samolotach
 - `flights` - dane o wylotach z danego lotniska per dzień.

 Wszystkie te źródła musimy pobrać, aby być w stanie wykonać całość warsztatu. W celu pobrania informacji, gdzie wymagany jest paramatr `airportId`, posłużymy się listą z pliku `airports.csv`.

 Przy wykonywaniu tego zadania możesz posłużyć się tym tokenem: `iKRsQ8vdqgT903o2vH1rsejOeQ0F7YC9TvutH6Wk`.

 ### Uwagi
 - Ze względów ćwiczeniowych, konstrukcja poszczególnych endpointów jest różna – w trakcie pracy dokładnie przyjrzyj się, w jaki sposób należy wykonać zapytanie, aby otrzymać odpowiedź.
 - Pamiętaj o dodaniu `sleep` pomiędzy poszczególnymi wywołaniami endpoint.
 - Limit wywołań API to 1000/min, zadbaj o nieprzekroczenie tego limitu – w przeciwnym wypadku będzie zwracany błąd 429.

 # Konfiguracja notatnika

 Tutaj zaimportuj wymagane biblioteki

import json
import pandas as pd
import requests
from time import sleep
from datetime import datetime, timedelta
from dateutil.relativedelta import relativedelta

 Tutaj zdefiniuj paramatry połączenia do API

API_KEY = "iKRsQ8vdqgT903o2vH1rsejOeQ0F7YC9TvutH6Wk"
API_URL = "https://api-datalab.coderslab.com/api/v2"

AUTH = {'authorization': API_KEY}

SLEEP = 60/950  # czas pomiędzy kolejnymi żądaniami w sekundach

 Tutaj wczytaj plik `airports.csv` i dostosuj do dalszych kroków w celu pobierania z kolejnych endpointów. Lista lotnisk jest dostępna w kolumnie `origin_airport_id`.

file_path = "C:/Users/User/OneDrive - AEI/Pulpit/Pliki_do_Portfolio_Lab/data/airports.csv"  #ścieżka do pliku
df = pd.read_csv(file_path)
df.head(5)

 # Pobieranie `Airport`
 Zapoznaj się z dokumentacją endpointu `airport`, a następnie pobierz dane dot. poszczególnych lotnisk. Wyniki tego kroku zapisz do ramki `airport_df`, a następnie zapisz do pliku `csv`.

 ### Wskazówki
 - Nie wszystkie lotniska dostępne w pliku `airports.csv`, są dostępne w endpoint. Zadbaj o odpowiednie obsłużenie takiej sytuacji,
 - Do skonwertowania wyników przydatna może okazać się metoda `Pandas` - [from_records](https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.from_records.html),
 - Artykuł LMS: `Python - analiza danych > Dzień 4 - API > Uwierzytelnianie`
 - Artykuł LMS: `Python - analiza danych > Przygotowanie do zjazdu 2`

 Tutaj pobierz dane z endpoint'u `airport`

# Zamień DataFrame na listę
lista = df.values.tolist()

# Rozpakowanie do listy jednowymiarowej
flat_list = [item for sublist in lista for item in sublist]
  
# Pusta lista na dane o lotniskach
airport_data = []

# Pobieranie danych dla każdego identyfikatora
for airport_id in flat_list:
    url = f"{API_URL}/airport/{airport_id}"
    response = requests.get(url, headers={'authorization': API_KEY})
    if response.status_code == 200:
        airport_data.append(response.json())
    else:
        print(f"Błąd pobierania danych dla lotniska o ID {airport_id}")

# Tworzenie ramki danych z wyników
airport_df = pd.DataFrame.from_records(airport_data)


airport_df

 ## Sprawdzenie
 Uruchom kod poniżej, aby sprawdzić czy ta część została poprawnie wykonana

airport_df_expected_shape = (97, 4)
assert airport_df_expected_shape == airport_df.shape

 Tutaj zapisz ramkę `airport_df` do pliku `airport_list.csv`

# Zapisz ramkę danych do pliku CSV
airport_df.to_csv('C:/Users/User/OneDrive - AEI/Pulpit/Pliki_do_Portfolio_Lab/data/raw/airport_list.csv', index=False)

 # Pobieranie `Weather`
 Zapoznaj się z dokumentacją endpotu `Weather`, następnie pobierz dane dotyczące zarejestrowanej pogody na poszczególnych lotniskach. Wyniki zapisz do ramki `weather_df`, a później do pliku `airport_weather.csv`.

 Wskazówki:
 - Ze względu na wolumen danych, które tutaj się pobiorą, odradzamy zapisywanie danych bezpośrednio do ramki. Rekomendujemy podejście podobne do tego z warsztatu na kursie `Python - analiza danych` - `Dzień 10 - Warsztat > Warsztat > Scrapowanie danych`, czyli stworzenie listy, a następnie przekonwertowanie jej w postać ramki.
 - Data początkowa danych to `2019-01-01`, zaś data końcowa to `2020-03-31`, czyli 15 miesięcy,
 - Ze względu na czas, jaki ten krok będzie się wykonywał, warto dodać w pętli instrukcję (lub kilka) `print`, aby monitorować przebieg wykonywania tego kroku.
 - Przy dodawaniu miesięcy do daty może przydać się metoda [relativedelta](https://www.geeksforgeeks.org/python-get-month-from-year-and-weekday/).

endpoint = "/airportWeather"
# Ustawienie dat początkowej i końcowej
start_date = datetime(2019, 1, 1)
end_date = datetime(2020, 3, 31)
# Lista, która będzie przechowywać dane
weather_data = []
current_date = start_date
# Pętla przez okres czasu
while current_date <= end_date:
    # Formatowanie daty
    formatted_date = current_date.strftime("%Y-%m-%d")
    # Tworzenie URL-a z odpowiednią datą
    url = f"{API_URL}{endpoint}?date={formatted_date}"
    # Wykonanie zapytania HTTP
    response = requests.get(url, headers={'authorization': API_KEY})
    # Sprawdzenie poprawności odpowiedzi
    if response.status_code == 200:
        data = response.json()
        weather_data.extend(data)
        print(f"Pobrano dane dla {formatted_date}")
    else:
        print(f"Błąd pobierania danych dla {formatted_date}")
    # Przejście do kolejnego miesiąca
    current_date += relativedelta(months=1)
    sleep(0.5)

# Przekształcenie słownika w ramkę danych
weather_df = pd.DataFrame.from_records(weather_data)

# Wyświetlenie ramki danych
weather_df

 ## Sprawdzenie
 Uruchom kod poniżej, aby sprawdzić, czy ta część została poprawnie wykonana

airport_weather_df_expected_shape = (46226, 33)
assert airport_weather_df_expected_shape == weather_df.shape

 ## Zapis do pliku
 Tutaj zapisz ramkę `weather_df` do pliku `airport_weather.csv` w katalogu `data/raw`

weather_df.to_csv('C:/Users/User/OneDrive - AEI/Pulpit/Pliki_do_Portfolio_Lab/data/raw/airport_weather.csv', index=False)

 # Pobranie `Aircraft`
 Zapoznaj się z dokumentacją endpointu `aircraft`, a następnie pobierz dane produkcyjne samolotów. Wyniki zapisz do ramki `aircraft_df`, a następnie zapisz do pliku `aircraft.csv`.


# Endpoint
endpoint = "/aircraft"

# Lista, która będzie przechowywać dane
aircraft_data = []

# Tworzenie URL-a
url = f"{API_URL}{endpoint}"

# Wykonanie zapytania HTTP
response = requests.get(url, headers={'authorization': API_KEY})

# Sprawdzenie poprawności odpowiedzi
if response.status_code == 200:
    data = response.json()
    # Zachowanie danych w liście
    aircraft_data.extend(data)
    print("Dane zostały pobrane i zapisane w liście.")
else:
    print("Błąd podczas pobierania danych")
sleep(0.01)

# Przekształcenie listy w ramkę danych
aircraft_df = pd.DataFrame.from_dict(aircraft_data)

# Wyświetlenie ramki danych
aircraft_df

 ## Sprawdzenie
 Uruchom kod poniżej, aby sprawdzić, czy ta część została poprawnie wykonana

aircraft_df_expected_shape = (7383, 3)
assert aircraft_df_expected_shape == aircraft_df.shape

 ## Zapis do pliku
 Tutaj zapisz ramkę `aircraft_df` do pliku `aircraft.csv` w katalogu `data/raw`

aircraft_df.to_csv('C:/Users/User/OneDrive - AEI/Pulpit/Pliki_do_Portfolio_Lab/data/raw/aircraft.csv', index=False)

 # Pobranie `Flight`
 Zapoznaj się z dokumentacją endpointu `flights`, następnie pobierz dane dotyczące ruchu lotniczego. Wyniki zapisz do ramki `flight_df`, a później do pliku `flight.csv`.

 Wskazówki:
 - Zwróć szczególną uwagę na konstrukcję endpointa,
 - Ze względu na wolumen danych, które tutaj się pobiorą, odradzamy zapisywanie danych bezpośrednio do ramki. Rekomendujemy podejście podobne do tego, z warsztatu na kursie `Python - analiza danych` - `Dzień 10 - Warsztat > Warsztat > Scrapowanie danych`,
 - Data początkowa danych to `2019-01-01`, zaś końcowa to `2020-03-31`, czyli 456 dni,
 - Ze względu na czas, jaki ten krok będzie się wykonywał, warto dodać w pętli instrukcję (lub kilka) `print`, aby monitorować przebieg wykonywania tego kroku,
 - W przypadku, gdy nie ma dostępnych danych dla danego lotniska, API zwraca kod [204](https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/204), w ten sposób możesz pominąć lotniska, dla których dane nie są dostępne,
 - Pobranie całości danych zajmuje dłuższą chwilę, zanim włączysz pętle dla wszystkich danych, sprawdź pobieranie danych dla jednego, dwóch lotnisk aby uniknąć frustracji.

# Otwarcie pliku csv z apirpots_id
file_folder= "C:/Users/User/OneDrive - AEI/Pulpit/Pliki_do_Portfolio_Lab/data/airports.csv"  #ścieżka do pliku
df_airports_id = pd.read_csv(file_folder)

# Zamień DataFrame na listę
origin_airport_ids = df_airports_id['origin_airport_id'].tolist()

endpoint_ft = '/flight'

# Lista do przechowywania danych o pogodzie na lotniskach
flight_data_list = []

# Data początkowa i końcowa
start_date = datetime(2019, 1, 1)

current_date = start_date
# Pobranie danych dla każdego lotniska
for airport_id in origin_airport_ids:
    # Pętla przez okres czasu
    current_date = start_date
    for i in range(15):
        # Formatowanie daty
        formatted_date = current_date.strftime("%Y-%m")
        # Tworzenie URL-a
        url = f"{API_URL}{endpoint_ft}?date={formatted_date}&airportId={airport_id}"
        # Wykonanie zapytania HTTP
        response = requests.get(url, headers={'authorization': API_KEY})
        # Sprawdzenie poprawności odpowiedzi
        if response.status_code == 200:
            data = response.json()
            if data:  # Sprawdzenie, czy dane istnieją
                flight_data_list.extend(data)
                print(f"Data downloaded for airport {airport_id} on {formatted_date}")
            else:
                print(f"No data available for the airport {airport_id} on {formatted_date}")
        elif response.status_code == 204:
            print(f"No data available for the airport {airport_id} on {formatted_date}")
            break
    # Przejście do kolejnej daty
        current_date += relativedelta(months=1)

# Przekształcenie słownika w ramkę danych
flight_df = pd.DataFrame.from_records(flight_data_list)

# Wyświetlenie ramki danych
flight_df

flight_df

 ## Sprawdzenie
 Uruchom kod poniżej, aby sprawdzić, czy ta część została poprawnie wykonana

flight_df_expected_shape = (1386120, 27)
assert flight_df_expected_shape == flight_df.shape

 ## Zapis do pliku
 Tutaj zapisz ramkę `flight_df` do pliku `flight.csv` w katalogu `data/raw`

flight_df.to_csv('C:/Users/User/OneDrive - AEI/Pulpit/Pliki_do_Portfolio_Lab/data/raw/flight.csv', index=False)

 # Podsumowanie
 W tym notatniku wykonaliśmy podstawowy krok w analizie danych - pozyskaliśmy je. Są gotowe do dalszej pracy, czyli możemy załadować je na bazę danych, a następnie zapoznać się z tym, jakie informacje ze sobą niosą. Kolejne notatniki będą służyły właśnie tym celom.

msg = "Wszystko wygląda OK :) Możesz przejść do kolejnego kroku."
print(msg)
