==> 02_Inicjowanie_bazy_danych.ipynb <==
 # Opis notatnika
 W tym notatniku skupimy się na stworzeniu dedykowanej bazy danych wraz ze strukturą tabel. Jednak w tym momencie nie będziemy jej uzupełniać danymi - tym zajmiemy się w kolejnej części.
 
 Ponownie wcielasz się w rolę Data Engineera, którego zadaniem jest dostarczenie klientowi bazy danych, która będzie używana do celów analityczno-raportowych. Bazy danych są podstawowym źródłem pracy dla Analityków Danych.
 
 Plik wraz ze strukturą bazy znajdziesz w folderze `sql/database_schema.sql`. 

 > W tym zadaniu konieczne jest posiadanie zainstalowanego `Postgres` na swoim komputrze. Jeżeli z jakichś powodów został on usunięty po kursie `Python - Analiza Danych`, należy ponownie dokonać instalacji.

 > Posiadając działającą instancję serwera, można stworzyć po prostu nową bazę np. `airlines`. Stworzenie bazy danych powinno odbyć się w notatniku poprzez stworzenie obiektu cursor i wywołaniu zapytań znajdujących się w pliku .sql.

 W trakcie rozwiązywania tego notatnika przydadzą się następujące materiały:
 - `Python - analiza danych -> Przygotowanie do kursu -> Podstawy SQL - Praca samodzielna -> Instalacja bazy danych`
 - `Python - analiza danych -> Moduł 1 -> Dzień 3 - PostgreSQL -> SQL i Python`
 - `Python - analiza danych -> Prework -> Podstawy SQL - Praca samodzielna -> Instalacja bazy danych -> Przygotowanie bazy danych`
 
Inicjację bazy danych wykonaj w notatniku, nie bezpośrednio na Postgresie.

 Tutaj zaimportuj potrzebne biblioteki

!pip install psycopg2-binary
import psycopg2

 ## Połączenie z bazą danych
 Tutaj uzupełnij konfigurację połączenia

username = 'postgres'
password = '9292'

host = 'localhost'
database = 'airlines'
port = None

Tutaj zdefiniuj zmienną `con` oraz `cursor`

conn = psycopg2.connect(
    dbname=database,
    user=username,
    password=password,
    host=host,
    port=port
)
cursor = conn.cursor()

# Sprawdź, czy połączenie działa
cur = conn.cursor()
cur.execute('SELECT version();')
print(cur.fetchone()[0])  # Wyświetl wersję PostgreSQL
cur.close()

 ## Wczytanie pliku `database_schema.sql`
 Z katalogu `sql` wczytaj plik `database_schema.sql`

with open("C:/Users/User/OneDrive - AEI/Pulpit/Pliki_do_Portfolio_Lab/sql/database_schema.sql", 'r') as schema_file:
    schema_sql = schema_file.read()

 W tym miejscu odpowiednio rozdziel zawartość pliku `database_schema.sql` na mniejsze kwerendy używając `;`

# Tabela aircraft

# Wykonaj zapytanie SQL
cur = conn.cursor()
cur.execute('''
    DROP TABLE IF EXISTS public.aircraft CASCADE;
    CREATE TABLE IF NOT EXISTS public.aircraft
    (
        id serial,
        manufacture_year bigint,
        tail_num text,
        number_of_seats double precision
    );
''')
conn.commit()


# Widok v_aircraft_blocker

# Wykonaj zapytanie SQL
cur = conn.cursor()
cur.execute('''
    -- BLOCKER TO FORCE USAGE if_exists in PANDAS (REMOVE IF OBSTACLE)
    CREATE OR REPLACE VIEW public.v_aircraft_blocker AS
    SELECT 1 FROM public.aircraft LIMIT 1;
''')
conn.commit()

# Tabela airport_weather

# Wykonaj zapytanie SQL
cur = conn.cursor()
cur.execute('''
    DROP TABLE IF EXISTS public.airport_weather CASCADE;
    CREATE TABLE IF NOT EXISTS public.airport_weather
    (
        id serial,
        station text,
        name text,
        date text,
        awnd double precision,
        pgtm double precision,
        prcp double precision,
        snow double precision,
        snwd double precision,
        tavg double precision,
        tmax double precision,
        tmin double precision,
        wdf2 double precision,
        wdf5 double precision,
        wsf2 double precision,
        wsf5 double precision,
        wt01 double precision,
        wt02 double precision,
        wt03 double precision,
        wt04 double precision,
        wt05 double precision,
        wt06 double precision,
        wt07 double precision,
        wt08 double precision,
        wt09 double precision,
        wesd double precision,
        wt10 double precision,
        psun double precision,
        tsun double precision,
        sn32 double precision,
        sx32 double precision,
        tobs double precision,
        wt11 double precision,
        wt18 double precision
    );

    -- BLOCKER TO FORCE USAGE if_exists in PANDAS (REMOVE IF OBSTACLE)
    CREATE OR REPLACE VIEW public.v_aircraft_weather_blocker AS
    SELECT 1 FROM public.airport_weather LIMIT 1;
''')
conn.commit()

# Tabela flight

# Wykonaj zapytanie SQL
cur = conn.cursor()
cur.execute('''
    DROP TABLE IF EXISTS public.flight CASCADE;
    CREATE TABLE IF NOT EXISTS public.flight
    (
        id serial,
        month bigint,
        day_of_month bigint,
        day_of_week bigint,
        op_unique_carrier text,
        tail_num text,
        op_carrier_fl_num bigint,
        origin_airport_id bigint,
        dest_airport_id bigint,
        crs_dep_time bigint,
        dep_time double precision,
        dep_delay_new double precision,
        dep_time_blk text,
        crs_arr_time bigint,
        arr_time double precision,
        arr_delay_new double precision,
        arr_time_blk text,
        cancelled double precision,
        crs_elapsed_time double precision,
        actual_elapsed_time double precision,
        distance double precision,
        distance_group bigint,
        carrier_delay double precision,
        weather_delay double precision,
        nas_delay double precision,
        security_delay double precision,
        late_aircraft_delay double precision,
        year bigint
    );

    -- BLOCKER TO FORCE USAGE if_exists in PANDAS (REMOVE IF OBSTACLE)
    CREATE OR REPLACE VIEW public.v_flight_blocker AS
    SELECT 1 FROM public.flight LIMIT 1;
''')
conn.commit()

# Tabela airport_list

# Wykonaj zapytanie SQL
cur = conn.cursor()
cur.execute('''
    DROP TABLE IF EXISTS public.airport_list CASCADE;
    CREATE TABLE IF NOT EXISTS public.airport_list
    (
        id serial,
        origin_airport_id bigint,
        display_airport_name text,
        origin_city_name text,
        name text
    );

    CREATE OR REPLACE VIEW public.v_airport_list_blocker AS
    SELECT 1 FROM public.airport_list LIMIT 1;
''')
conn.commit()

 W tym miejscu wykonaj każdą z kwerend, aby zainicjować strukturę bazy danych

 Zatwierdznie wszystkich operacji na bazie, czyli stworzenie tabel

 ### Sprawdzenie
 Uruchom kod poniżej, aby sprawdzić, czy ta część została poprawnie wykonana

def check_if_table_exists(table_name):
    msg = f"Sprawdzam czy istnieje tabela {table_name}"
    print(msg)

    query = f"select 1 from {table_name}"
    # jeżeli tabela nie istnieje, ten krok zwróci wyjątek
    cursor.execute(query)
    print('OK!')


tables_to_test = [
    'aircraft',
    'airport_weather',
    'flight',
    'airport_list'
]

for table in tables_to_test:
    check_if_table_exists(table)

conn.close()
msg = "Wszystko wygląda OK :) Możesz przejść do kolejnego zadania."
print(msg)

 # Podsumowanie
 Za pomocą tego notatnika została zbudowana w sposób automatyczny nasza docelowa baza danych. Dzięki temu nie musimy się już martwić o jej ręczną przebudowę - w ramach potrzeby wystarczy włączyć notatnik.
 Wykonując kolejny notatnik, sprawimy, że w tabelach pojawią się również dane potrzebne do przeprowadzenia późniejszej analizy oraz wykorzystywane w raportowaniu.
